% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSetRatio.R
\name{calculate_geneSetRatio}
\alias{calculate_geneSetRatio}
\title{Calculate geneSetRatio per gene per enrichment contrast}
\usage{
calculate_geneSetRatio(enrichment_results, gene_overview_df)
}
\arguments{
\item{enrichment_results}{list of enrichment results}

\item{gene_overview_df}{dataframe with gene-wise information}
}
\value{
numerical vector of gene set ratios
}
\description{
Get a percentage of genesets the specific gene is included
}
\examples{
calculate_geneSetRatio(
list(
  A = goatea::example_enrichment, 
  B = goatea::example_enrichment
), 
goatea::example_genes_overview)
}
