% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorify.R
\name{colorify}
\alias{colorify}
\title{Create and/or modify color/gradient palettes}
\usage{
colorify(
  n = NULL,
  colors = character(0),
  n_gradient = n,
  h = 1,
  s = 1,
  l = 1,
  r = 1,
  g = 1,
  b = 1,
  alpha = 1,
  seed = 1L,
  rev = FALSE,
  plot = FALSE
)
}
\arguments{
\item{n}{default: NULL, else integer, amount of colors to create, if palette selected and more colors requested they will be generated}

\item{colors}{if integer, generate that many theoretically maximally different colors, select palette by name (options: see grDevices::palette.pals(), "maus"), or vector of R color names or color hexcodes}

\item{n_gradient}{default: n, else integer, amount of colors to output as gradient, after completing palette for n colors}

\item{h}{hue factor, default: 1, multiply values by factor, proportional scaling from base value of 1}

\item{s}{saturation factor, default: 1, multiply values by factor, proportional scaling from base value of 1}

\item{l}{lightness/brightness factor, default: 1, multiply values by factor, proportional scaling from base value of 1}

\item{r}{red factor, default: 1, multiply values by factor, proportional scaling from base value of 1}

\item{g}{green factor, default: 1, multiply values by factor, proportional scaling from base value of 1}

\item{b}{blue factor, default: 1, multiply values by factor, proportional scaling from base value of 1}

\item{alpha}{numeric, sets color alpha values}

\item{seed}{integer, set seed for generation of colors}

\item{rev}{default: FALSE, if TRUE, reverse order of colors}

\item{plot}{default: FALSE, if TRUE plot pie chart of color palette}
}
\value{
vector of color hexcodes
}
\description{
Note for colorblind use: "Okabe-Ito"
TODO add description
}
\details{
Either generate theoretically maximally different colors, select an available R grDevices palette and/or modify the colors of the given gradient/palette
}
\examples{
## if parameters identical, change seed to change generation
colorify(10, plot = TRUE, seed = 1)
colorify(10, plot = TRUE, seed = 42)
## set colors, generate additional up to n
colorify(colors = c("red", "white", "blue"), n = 5, plot = TRUE)
## create gradients
colorify(colors = c("orange", "red", "white", "blue", "orange"), n_gradient = 100, plot = TRUE)

## viridis gradient, lighten and saturate, darken
colorify(colors = "viridis", n = 100, plot = TRUE)
colorify(colors = "viridis", n = 10, plot = TRUE, l = 1.2, s = 10)
colorify(colors = "viridis", n = 10, plot = TRUE, l = .9)

TODO rgb, hsl examples
TODO all parameter examples
}
