% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorify.R
\name{colorify_map}
\alias{colorify_map}
\title{Colorify colorRamp between colors mapping to breakpoint values}
\usage{
colorify_map(colors, breakpoints, ...)
}
\arguments{
\item{colors}{hexcolor character vector}

\item{breakpoints}{numeric vector matching colors per value}

\item{...}{to pass arguments to grDevices::colorRamp}
}
\value{
function with colors and breaks attributes, can be called as function(c(values)) to return hexcolorcodes
}
\description{
Note that breakpoints and colors will be ordered ascendingly by breakpoints values
}
