% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{colorscale}
\alias{colorscale}
\title{Colorscale for given values and color breaks}
\usage{
colorscale(values, pos_color = "red", neg_color = "blue", zero_color = "white")
}
\arguments{
\item{values}{Values to create colorscale for}

\item{pos_color}{Named or hex color for 'positive' colors}

\item{neg_color}{Named or hex color for 'negative' colors}

\item{zero_color}{Named or hex color for 'zero'/'neutral' colors}
}
\value{
color function
}
\description{
Colorscale uses circlize::colorRamp2
}
