% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date2gene}
\alias{date2gene}
\title{Fix Excel formatting genes to dates back to gene names.}
\usage{
date2gene(gene_names)
}
\arguments{
\item{gene_names}{gene names (where some are dates) in character vector}
}
\value{
gene names in character vector
}
\description{
Fix Excel formatting genes to dates back to gene names.
}
