% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorify.R
\name{display_palettes}
\alias{display_palettes}
\title{Display R grDevices palettes}
\usage{
display_palettes(n = 10, i_palettes = 1:131, border = FALSE)
}
\arguments{
\item{n}{integer, amount of colors to display}

\item{i_palettes}{index range as numerical vector for choosing palettes, any i over maximum amount of palettes are removed}

\item{border}{default: FALSE, if TRUE show color rectangle borders}
}
\value{
named vector with source and name of palettes, 'hcl' for grDevices::hcl.pals() and 'pal' for grDevices::palette.pals()
}
\description{
Use colorify() to select and modify the palettes, see its documentation.
Note that discrete palettes with maximum n colors will be repeated in plotting.
}
\examples{
display_palettes()
display_palettes(i_palettes = 50:75)
display_palettes(i_palettes = c(1,5,10,20,40,100,119)
display_palettes(n = 100, i_palettes = 1:10)
display_palettes(n = 10, i_palettes = 1:10, border = TRUE)
}
