% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_Colameo_MS}
\alias{example_Colameo_MS}
\title{Example Colameo MS data}
\format{
A data frame with columns gene, symbol, effectsize, pvalue.
}
\source{
Colameo et al. 2021 (PMID: 34396684)
}
\usage{
example_Colameo_MS
}
\description{
Mass spectrometry genelist from Colameo et al. 2021.
}
\keyword{datasets}
