% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_Colameo_RNA}
\alias{example_Colameo_RNA}
\title{Example Colameo RNA data}
\format{
A data frame with columns gene, symbol, effectsize, pvalue.
}
\source{
Colameo et al. 2021 (PMID: 34396684)
}
\usage{
example_Colameo_RNA
}
\description{
RNA-seq genelist from Colameo et al. 2021.
}
\keyword{datasets}
