% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_enrichment}
\alias{example_enrichment}
\title{An example enrichment}
\format{
\subsection{\code{enrichment}}{

A data frame with 10 rows and 17 columns:
\describe{
\item{source}{origin}
\item{source_version}{org.Xx.eg.db}
\item{id}{DB.001}
\item{name}{geneset name 1, geneset name 2}
\item{parent_id}{DB.010, DB.020}
\item{ngenes_input}{10, 30}
\item{ngenes}{10, 30}
\item{genes}{10000, 10001}
\item{ngenes_signif}{10, 30}
\item{score_type}{effectsize}
\item{pvalue}{0.05, 1}
\item{zscore}{-Inf, 0, Inf}
...
}
}
}
\source{
generated with data-raw/example_data.R
}
\usage{
example_enrichment
}
\description{
An example enrichment
}
\keyword{datasets}
