% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genelist}
\alias{example_genelist}
\title{An example genelist}
\format{
\subsection{\code{genelist}}{

A data frame with 100 rows and 4 columns:
\describe{
\item{symbol}{Gene_1, Gene_2}
\item{gene}{10000, 10001}
\item{pvalue}{0.05, 1}
\item{effectsize}{2.5, 0}
...
}
}
}
\source{
generated with \code{data-raw/example_data.R}
}
\description{
A simulated example of a genelist for testing or demonstration purposes.
}
