% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genes_overview}
\alias{example_genes_overview}
\title{An example genes overview}
\format{
\subsection{\verb{genes overview}}{

A data frame with 100 rows and ~11 columns:
\describe{
\item{gene}{10000, 10001}
\item{symbol}{Gene_1, Gene_2}
\item{sample_efsi}{2.5, 0}
\item{sample_pval}{0.05, 1}
\item{sample_perc}{0, 100}
\item{genelist_overlap}{"", "A", "B", "AB"}
\item{sample_geneSetRatio}{0, 50, 100}
...
}
}
}
\source{
generated with data-raw/example_data.R
}
\usage{
example_genes_overview
}
\description{
An example genes overview
}
\keyword{datasets}
