% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_genesets}
\alias{example_genesets}
\title{Example genesets}
\format{
\subsection{\code{genesets}}{

A data frame with 100 rows and 4 columns:
\describe{
\item{source}{origin}
\item{source_version}{org.Xx.eg.db}
\item{id}{DB.001}
\item{name}{geneset name 1, geneset name 2}
\item{parent_id}{DB.010, DB.020}
\item{genes}{10000, 10001}
\item{ngenes}{10, 30}
...
}
}
}
\source{
generated with data-raw/example_data.R
}
\usage{
example_genesets
}
\description{
Example genesets
}
\keyword{datasets}
