% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_ppi_data}
\alias{example_ppi_data}
\title{An example ppi data}
\format{
\subsection{\verb{ppi data}}{

A data frame with 15 rows and 5 columns:
\describe{
\item{from_symbol}{gene_A, gene_B}
\item{to_symbol}{gene_A, gene_B}
\item{combined_score}{0, 1000}
\item{from}{gene_A_ID, gene_B_ID}
\item{to}{gene_A_ID, gene_B_ID}
...
}
}
}
\source{
generated with \code{data-raw/example_data.R}
}
\description{
A simulated example of a ppi dataframe for testing or demonstration purposes.
}
