% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_extension}
\alias{file_extension}
\title{Get file extension}
\usage{
file_extension(x)
}
\arguments{
\item{x}{string filepath}
}
\value{
string file extension
}
\description{
Get file extension
}
\examples{
file_extension(file.path(getwd(), 'filename.txt'))
}
