% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_enrichment.R
\name{filter_enrichment}
\alias{filter_enrichment}
\title{Filter enrichment}
\usage{
filter_enrichment(
  df,
  genes_input = "",
  genes_any_all = c("any", "all"),
  terms_query = "",
  terms_query_all_any = c("any", "all"),
  terms_antiquery = "",
  terms_antiquery_all_any = c("any", "all"),
  min_ngenes = 0,
  min_ngenes_input = 0,
  min_ngenes_signif = 0,
  min_abs_zscore = 0,
  max_pvalue_adjust = 1
)
}
\arguments{
\item{df}{enrichment output dataframe}

\item{genes_input}{default: UI input/character vector of genes to select df terms for}

\item{terms_query}{dfeault: UI input/character vector of keywords to match (grepl) term names}

\item{terms_query_all_any}{default: 'any', else 'all', defines if terms should match any or all of the query keywords given}

\item{terms_antiquery}{dfeault: UI input/character vector of keywords to NOT match (grepl) term names}

\item{terms_antiquery_all_any}{default: 'any', else 'all', defines if terms should NOT match any or all of the query keywords given}

\item{min_ngenes}{default: 0, set higher to filter terms with less n genes}

\item{min_ngenes_signif}{default: 0, set higher to filter terms with less n significant genes}

\item{min_abs_zscore}{default: 0, set higher to filter terms with less absolute zscore}

\item{max_pvalue_adjust}{default: 1, set lower to filter terms with lower adjusted p-value for multiple correction}

\item{genes_type}{default: 'any', else 'all', use to define to take only specific terms containing any or all associated genes}
}
\description{
Search and filter and sort or summarize (compiled) enrichment output.
}
