% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_terms_by_keywords}
\alias{get_terms_by_keywords}
\title{Get term names by searching with (partial) keywords}
\usage{
get_terms_by_keywords(patterns, terms, pos_neg = "pos", all_any = "all")
}
\arguments{
\item{patterns}{keywords to match (grepl) term names}

\item{terms}{character vector to be grepl searched}

\item{pos_neg}{return positive matches or negate matches}

\item{all_any}{need all or any patterns to match search terms}
}
\description{
Get term names by searching with (partial) keywords
}
