% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/goatea-package.R
\docType{package}
\name{goatea-package}
\alias{goatea}
\alias{goatea-package}
\title{goatea: Interactive Exploration of GSEA by the GOAT Method}
\description{
Geneset Ordinal Association Test Enrichment Analysis (GOATEA) provides a 'Shiny' interface with interactive visualizations and utility functions for performing and exploring gene set enrichment analysis using the 'GOAT' package. 'GOATEA' is designed to support large-scale and user-friendly enrichment workflows across multiple gene lists and comparisons, with flexible plotting and output options. Visualizations pre-enrichment include interactive 'Volcano' and 'UpSet' (overlap) plots. Visualizations post-enrichment include interactive geneset dotplot, geneset treeplot, gene-effectsize heatmap, gene-geneset heatmap and 'STRING' database of protein-protein-interactions network graph. 'GOAT' reference: Frank Koopmans (2024) \doi{ https://doi.org/10.1038/s42003-024-06454-5}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/mauritsunkel/goatea}
  \item Report bugs at \url{https://github.com/mauritsunkel/goatea/issues}
}

}
\author{
\strong{Maintainer}: Maurits Unkel \email{mauritsunkel@gmail.com} (\href{https://orcid.org/0000-0003-1920-6001}{ORCID})

}
\keyword{internal}
