% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{hexcolor2rgba}
\alias{hexcolor2rgba}
\title{Hex code colors to rgba format}
\usage{
hexcolor2rgba(hexcolors, alpha = NULL)
}
\arguments{
\item{hexcolors}{character (vector), hexcode colors (e.g. #FFFFFF)}

\item{alpha}{numeric in range \code{[0-1]}, default: NULL to use full opacity or given opacity (AA) in hexcolors (#RRGGBBAA)}
}
\value{
colors in rgba format
}
\description{
Hex code colors to rgba format
}
\examples{
colors <- colorify(5)
hexcolor2rgba(colors)
hexcolor2rgba(colors, alpha = .5)
colors <- gsub('FF$', 75, colors)
hexcolor2rgba(colors)
hexcolor2rgba(colors, alpha = .5)
}
