% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{\%||\%}
\alias{\%||\%}
\title{Null-coalescing operator}
\usage{
lhs \%||\% rhs
}
\arguments{
\item{lhs}{A value that may be NULL.}

\item{rhs}{A default value to return if lhs is NULL.}
}
\value{
The value of \code{lhs} if it is not NULL, otherwise \code{rhs}.
}
\description{
See \code{\link[rlang]{\%||\%}} for details.
}
\keyword{internal}
