% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorify.R
\name{palette_name_mapping}
\alias{palette_name_mapping}
\title{Palette original name mapping}
\usage{
palette_name_mapping(palette)
}
\arguments{
\item{palette}{string: name of palette, will be lower()ed and stripped of whitespace}
}
\value{
original palette name
}
\description{
All ColorBrewer palettes overlap with grDevices palettes
Viridis palettes, except "Magma", overlap with grDevices palettes
}
\examples{
palette_name_mapping("dark2") # "Dark 2"
}
