% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_genelists_overlap.R
\name{plot_genelists_overlap_upset}
\alias{plot_genelists_overlap_upset}
\title{Visualize genelists gene overlap in UpSet plot or Venn diagram}
\usage{
plot_genelists_overlap_upset(
  genelists,
  grayscale_colors = TRUE,
  empty_intersections = TRUE
)
}
\arguments{
\item{genelists}{UI value/list of tibbles/dataframes}

\item{grayscale_colors}{default: UI valaue/FALSE, else TRUE, for colorful or grayscale degrees in UpSet plot}

\item{empty_intersections}{default: UI/value/TRUE, if FALSE, shouldn't print intersects with 0 overlaps, error in UpSetR package}
}
\value{
plot to render
}
\description{
Visualize genelists gene overlap in UpSet plot or Venn diagram
}
