% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_termtree.R
\name{plot_termtree}
\alias{plot_termtree}
\title{Plot semantic similarity termtree}
\usage{
plot_termtree(enrichment_result, Nterms = 50, Nwords = 5)
}
\arguments{
\item{enrichment_result}{GOAT enrichment result dataframe}

\item{Nterms}{integer, default: 50, sets amount of terms to plot}

\item{Nwords}{integer, default: 5, sets N summarized words per cluster}
}
\value{
ggtree/gg/ggplot object
}
\description{
Plot semantic similarity termtree
}
