% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_area_input}
\alias{process_area_input}
\title{Process Shiny areaInput string}
\usage{
process_area_input(area_input)
}
\arguments{
\item{area_input}{Shiny areaInput string}
}
\value{
string
}
\description{
Process Shiny areaInput string
}
