% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_string_input}
\alias{process_string_input}
\title{Process Shiny area input string}
\usage{
process_string_input(string_input)
}
\arguments{
\item{string_input}{siny string}
}
\value{
processed string - no whitespace, enters, only letters and numbers
}
\description{
Process Shiny area input string
}
\examples{
process_string_input("test string \n")
}
