% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_write_merged_enrichments}
\alias{process_write_merged_enrichments}
\title{Helper function to process and write combined genesets overview and genes overview output}
\usage{
process_write_merged_enrichments(
  merged_enrichment,
  output_folder,
  filename,
  top_n = NULL
)
}
\arguments{
\item{merged_enrichment}{row binded enrichment results dataframe/tibble}

\item{output_folder}{base output folder path to write to}

\item{top_n}{n top results per genelist per source based on genesets adjusted pvalue}
}
\description{
Helper function to process and write combined genesets overview and genes overview output
}
