% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{rename_gene_overview}
\alias{rename_gene_overview}
\title{Rename the gene overview}
\usage{
rename_gene_overview(names, genes_overview)
}
\arguments{
\item{names}{names to rename gene overview}

\item{genes_overview}{UI given genes overview dataframe (rv_genelists_overlap$gene_overview)}
}
\value{
genes overview renamed
}
\description{
Rename the gene overview
}
