% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_genelists_overlap.R
\name{run_genelists_overlap}
\alias{run_genelists_overlap}
\title{Create gene overview through overlapping genelists information by overlapping significant genes}
\usage{
run_genelists_overlap(
  genelists,
  annotate_genes = TRUE,
  annotation_organism = "Hs"
)
}
\arguments{
\item{genelists}{UI value/list of tibbles/dataframes}

\item{annotate_genes}{default: UI value/TRUE, else FALSE to not add small gene descriptive annotations}

\item{annotation_organism}{default: UI value/"Hs" (Homo Sapiens), otherwise: "Mm" (Mus Musculus)}
}
\value{
tibble/dataframe with (annotated) genes and p-value/effectsize info for each genelist, concluding with overlapping genelists by significant genes
}
\description{
Create gene overview through overlapping genelists information by overlapping significant genes
}
