% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scale_values_between}
\alias{scale_values_between}
\title{Scale values between given min/max}
\usage{
scale_values_between(
  values,
  old_min = min(values),
  old_max = max(values),
  new_min = 0,
  new_max = 100
)
}
\arguments{
\item{values}{numeric (vector)}

\item{old_min}{numeric, default: min(values), else set as current expected minimum of values}

\item{old_max}{numeric, default: max(values), else set as current expected maximum of values}

\item{new_min}{numeric, default: 0, else set to wanted new minimum value}

\item{new_max}{numeric, default: 100, else set to wanted new maximum value}
}
\value{
scaled numeric values
}
\description{
Scale values between given min/max
}
\examples{
scale_values_between(c(1,3,1,4,1,6,1,6,5,7))
}
