% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server for goatea package}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{Shiny input elements handling}

\item{output}{Shiny input elements handling}

\item{session}{Shiny handling reactivity in app}
}
\description{
Server for goatea package
}
