% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{set_base_folder}
\alias{set_base_folder}
\title{Set base folder}
\usage{
set_base_folder(folder_path = NULL)
}
\arguments{
\item{folder_path}{character, default NULL, else existing directory}
}
\value{
character folder path
}
\description{
sets given folder path if provided
else checks in order:
\itemize{
\item path.expand("~") (tilde (~) expands to HOME folder path)
\item Sys.getenv("R_USER") (set on R session start)
\item Sys.getenv("USERPROFILE") (Windows specific)
}
}
\examples{
set_base_folder()
}
