% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_hovertip}
\alias{wrap_hovertip}
\title{Wrap Shiny UI element with a hoverable tooltip contained in html div tags}
\usage{
wrap_hovertip(ui_element, hovertip)
}
\arguments{
\item{ui_element}{Shiny UI element to wrap with hovertext}

\item{hovertip}{text that will show as hover popup}
}
\value{
tags$div element around given Shiny UI element
}
\description{
Wrap Shiny UI element with a hoverable tooltip contained in html div tags
}
\examples{
wrap_hovertip(shiny::actionButton('id_example', 'example'), 'example')
}
