% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_geneSetRatio.R
\name{calculate_geneSetRatio}
\alias{calculate_geneSetRatio}
\title{Calculate geneSetRatio per gene per enrichment contrast}
\usage{
calculate_geneSetRatio(enrichment_results, gene_overview_df)
}
\arguments{
\item{enrichment_results}{list of enrichment results}

\item{gene_overview_df}{dataframe with gene-wise information}
}
\description{
Get a percentage of genesets the specific gene is included
}
