% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colorify.R
\name{colorify_map}
\alias{colorify_map}
\title{Colorify colorRamp between colors mapping to breakpoint values}
\usage{
colorify_map(colors, breakpoints, ...)
}
\arguments{
\item{colors}{hexcolor character vector}

\item{breakpoints}{numeric vector matching colors per value}

\item{...}{to pass arguments to grDevices::colorRamp}

\item{n_breakpoints}{optional, default: 1000, else non-negative numeric value for amount of breakpoints
, n_breakpoints = 1000}
}
\value{
function with colors and breaks attributes, can be called as function(c(values)) to return hexcolorcodes
}
\description{
Note that breakpoints and colors will be ordered ascendingly by breakpoints values
}
\examples{
map_f <- colorify_map(colors = c("white", "blue", "red"), breakpoints = c(0, 10, -5))
colorify(colors=map_f(c(-1,0,2,3)), plot=T)
}
