% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_gene_annotation}
\alias{get_gene_annotation}
\title{Get gene annotation for mouse}
\usage{
get_gene_annotation(gene_symbols, organism = "Hs")
}
\arguments{
\item{gene_symbols}{character vector of ensembl genes 109 symbols}

\item{organism}{default "Hs" (Homo Sapiens), otherwise: "Mm" (Mus Musculus)}
}
\description{
Can extend functino to get gene annotations for other organisms within annotables package.
}
