% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{goatea_server}
\alias{goatea_server}
\title{Server for goatea package}
\usage{
goatea_server(input, output, session, css_colors)
}
\arguments{
\item{input}{Shiny input elements handling}

\item{output}{Shiny input elements handling}

\item{session}{Shiny handling reactivity in app}

\item{css_colors}{see app.R, user set manual colors for the GOATEA UI}
}
\value{
Shiny server function
}
\description{
Server for goatea package
}
