% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_EnhancedVolcano.R
\name{plot_EnhancedVolcano}
\alias{plot_EnhancedVolcano}
\title{Plot EnhancedVolcano}
\usage{
plot_EnhancedVolcano(
  genelist,
  effectsize_threshold = 1,
  pvalue_threshold = 0.05,
  interactive = FALSE
)
}
\arguments{
\item{genelist}{UI value/list of tibbles/dataframes}

\item{effectsize_threshold}{numeric, default: 1, threshold for showing significance on effectsize axis}

\item{pvalue_threshold}{numeric, default: 0.05, threshold for showing significance on pvalue axis}

\item{interactive}{default: FALSE, else TRUE}
}
\value{
plotly or ggplot2 object
}
\description{
Plot EnhancedVolcano
}
