% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_termtree.R
\name{plot_termtree}
\alias{plot_termtree}
\title{Plot semantic similarity termtree}
\usage{
plot_termtree(
  genelist,
  map_organism = 9606,
  effectsize_threshold = 1,
  Nterms = NA,
  Nwords = 5,
  Nclusters = 1
)
}
\arguments{
\item{genelist}{GOAT current genelist from selected enrichment sample}

\item{map_organism}{integer, default: 9606 (human) - input organism ID that will be mapped to org.Xx.eg.db}

\item{effectsize_threshold}{numerical, default: 1 - genelist effectsize threshold}

\item{Nterms}{integer, default: NA to plat all terms, integer sets amount of terms to plot}

\item{Nwords}{integer, default: 5, sets N summarized words per cluster}

\item{Nclusters}{integer, default: 1, sets N clusters of terms}
}
\value{
ggtree/gg/ggplot object
}
\description{
Plot semantic similarity termtree
}
\examples{
plot_termtree(genelist <- get(load(system.file("extdata", "example_genelist.rda", package = "goatea"))))
}
