% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_termtree.R
\name{plot_termtree}
\alias{plot_termtree}
\title{Plot semantic similarity termtree}
\usage{
plot_termtree(enrichment, Nterms = NA, Nwords = 5, Nclusters = 1)
}
\arguments{
\item{enrichment}{GOAT enrichment result dataframe}

\item{Nterms}{integer, default: NA to plat all terms, integer sets amount of terms to plot}

\item{Nwords}{integer, default: 5, sets N summarized words per cluster}
}
\value{
ggtree/gg/ggplot object
}
\description{
Plot semantic similarity termtree
}
