% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{process_string_input}
\alias{process_string_input}
\title{Process Shiny areaInput string}
\usage{
process_string_input(string_input)
}
\arguments{
\item{string_input}{string}
}
\value{
processed string without enters, whitespace and only containing letters and numbers
}
\description{
Process Shiny areaInput string
}
