% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_genelist.R
\name{read_genelist}
\alias{read_genelist}
\title{Read differential gene expression results .csvs/.xlsx to comparisons format}
\usage{
read_genelist(filepaths)
}
\arguments{
\item{filepaths}{named character vector of full filepathss of DE.csvs/DE.xlsx (output by enrichment_analysis())}
}
\description{
Read differential gene expression results .csvs/.xlsx to comparisons format
}
