% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_validate_genelist.R
\name{read_validate_genelist}
\alias{read_validate_genelist}
\title{Read and validate a table with genes (that should be tested in overrepresentation-analysis) for compatibility with this R package#'}
\usage{
read_validate_genelist(
  file,
  remove_NA_ids = TRUE,
  remove_duplicated = TRUE,
  remove_Rik_genes = TRUE,
  remove_Gm_genes = TRUE,
  keep_maxN_genes = TRUE
)
}
\arguments{
\item{file}{full filepath to gene tibble in .csvs/.xlsx/.tsv}

\item{remove_NA_ids}{boolean, default TRUE, if non-integers in gene column, remove}

\item{remove_duplicated}{boolean, default TRUE, removes duplicated gene symbols/ids}

\item{remove_Rik_genes}{boolean, default TRUE, grepl("Rik$") search and remove Riken non-canonical mouse genes}

\item{remove_Gm_genes}{boolean, default TRUE, grepl("^Gm") search and remove Gm non-canonical mouse genes}

\item{keep_maxN_genes}{boolean, default TRUE, filter down by pvalue to max n genes allowed by goat (max(goat::goat_nulldistributions$N))}
}
\description{
Read and validate a table with genes (that should be tested in overrepresentation-analysis) for compatibility with this R package#'
}
