% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genelists_overlap.R
\name{run_genelists_overlap}
\alias{run_genelists_overlap}
\title{Create gene overview through overlapping genelists information by overlapping significant genes}
\usage{
run_genelists_overlap(genelists)
}
\arguments{
\item{genelists}{UI value/list of tibbles/dataframes}
}
\value{
tibble/dataframe with (annotated) genes and p-value/effectsize info for each genelist, concluding with overlapping genelists by significant genes
}
\description{
Create gene overview through overlapping genelists information by overlapping significant genes
}
\examples{
run_genelists_overlap(list(
  A = get(load(system.file("extdata", "example_genelist.rda", package = "goatea"))), 
  B = get(load(system.file("extdata", "example_genelist.rda", package = "goatea")))
))
}
