% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wrap_hovertip}
\alias{wrap_hovertip}
\title{Wrap Shiny UI element with a hoverable tooltip contained in html div tags}
\usage{
wrap_hovertip(text, ui_element)
}
\arguments{
\item{text}{}

\item{ui_element}{}
}
\description{
Wrap Shiny UI element with a hoverable tooltip contained in html div tags
}
